requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0
cryptography>=42.0.0
anthropic>=0.20.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
